TK-434: PR corrections + problem with the error message.
