+++ tumorotek-webapp/src/main/webapp/WEB-INF/internationalization/zk-label.properties
@@ -503,7 +503,6 @@ ficheProdDerive.error.parent.prelevement=the parent sampling :
 ficheProdDerive.error.parent.echantillon=the parent sample:
 ficheProdDerive.error.parent.derive=the parent byproduct:
 ficheProdDerive.error.type=Byproduct type is mandatory
-ficheProdDerive.error.quantite=The quantity used is mandatory
 ficheProdDerive.warning.quantite=Attention: Quantity not provided. Do you wish to proceed, knowing that automatic destocking by the parent cannot be managed?
 prodDerive.selection.multiple=Byproducts selection
 Champ.ProdDerive.Reservation.Debut=Starting date
