+++ tumorotek-webapp/src/main/webapp/WEB-INF/internationalization/zk-label_fr.properties
@@ -521,7 +521,6 @@ ficheProdDerive.error.parent.prelevement= le prélèvement parent :
 ficheProdDerive.error.parent.echantillon= l'échantillon parent : 
 ficheProdDerive.error.parent.derive= le produit dérivé parent : 
 ficheProdDerive.error.type=Type du produit dérivé obligatoire
-ficheProdDerive.error.quantite=La quantité utilisée est obligatoire
 ficheProdDerive.warning.quantite=Attention! quantité non renseignée. Souhaitez vous poursuivre sachant que le déstockage automatique du parent ne pourra pas être géré ?
