+++ tumorotek-webapp/src/main/webapp/zuls/prodderive/FicheMultiProdDerive.zul
@@ -204,11 +204,10 @@
 						<calendarbox id="dateTransfoCalBox" value="@{fwinMultiProdDerive$composer.object.dateTransformation,
 			        			save-when='validate.onClick, create.onClick, self.onBlur, interfacage.onClick'}" />
 							
-						<!-- Quantité utilisé dans la transformation -->
+						<!-- Quantité utilisée dans la transformation -->
 						<hbox>
 							<label id="transfoQuantiteLabel" value="${c:l('ficheProdDerive.transformation.quantité')}" sclass="formLabel"/>
-							<label id="requiredtransfoQuantiteLabel" value="*" sclass="requiredMark"
-								   visible="true"/>
+							<label id="requiredtransfoQuantiteLabel" value="*" sclass="requiredMark" />
 						</hbox>
 
 						<div id="transfoQuantiteDiv">
